+++ CHANGELOG.md
@@ -32,10 +32,15 @@ this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.htm
 
 - Bug where the `--ont` flag resulted in mykrobe predict crashing if the panel
   contained any presence/absence genes [[#123][123]].
+- Bug reporting variants that are "amino acid to any other amino acid" changes,
+  eg `katG_S315X`. If gene is on reverse strand, then the wrong new amino
+  acid change was being reported [[#127][127]].
+
 
 [113]: https://github.com/Mykrobe-tools/mykrobe/issues/113
 [114]: https://github.com/Mykrobe-tools/mykrobe/issues/114
 [123]: https://github.com/Mykrobe-tools/mykrobe/issues/123
+[127]: https://github.com/Mykrobe-tools/mykrobe/issues/127
 [Unreleased]: https://github.com/Mykrobe-tools/mykrobe/compare/v0.9.0...HEAD
 [mkdtemp]: https://docs.python.org/3.6/library/tempfile.html#tempfile.mkdtemp